requirements.txt
matplotlib==3.6.3
pytube==12.1.2
requests==2.28.2
python-dotenv~=0.21.1
chardet==5.1.0